Move channel and codec related files to common
This is in preparation for implementing channel and plugin support
for macOS desktop builds.
