+++ shell/platform/darwin/common/FlutterChannels.mm
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/platform/darwin/ios/framework/Headers/FlutterChannels.h"
+#include "flutter/shell/platform/darwin/common/FlutterChannels.h"
 
 #pragma mark - Basic message channel
