+++ shell/platform/darwin/common/FlutterCodecs.mm
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/platform/darwin/ios/framework/Headers/FlutterCodecs.h"
+#include "flutter/shell/platform/darwin/common/FlutterCodecs.h"
 
 @implementation FlutterBinaryCodec
 + (instancetype)sharedInstance {
