+++ shell/platform/darwin/common/flutter_codecs_unittest.mm
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/platform/darwin/ios/framework/Headers/FlutterCodecs.h"
+#include "flutter/shell/platform/darwin/common/FlutterCodecs.h"
 #include "gtest/gtest.h"
 
 TEST(FlutterStringCodec, CanEncodeAndDecodeNil) {
