+++ shell/platform/darwin/ios/BUILD.gn
@@ -26,16 +26,12 @@ shared_library("flutter_framework_dylib") {
     "framework/Headers/FlutterViewController.h",
     "framework/Headers/FlutterNavigationController.h",
     "framework/Source/FlutterAppDelegate.mm",
-    "framework/Source/FlutterChannels.mm",
-    "framework/Source/FlutterCodecs.mm",
     "framework/Source/FlutterDartProject.mm",
     "framework/Source/FlutterDartProject_Internal.h",
     "framework/Source/FlutterDartSource.h",
     "framework/Source/FlutterDartSource.mm",
     "framework/Source/FlutterPlatformPlugin.h",
     "framework/Source/FlutterPlatformPlugin.mm",
-    "framework/Source/FlutterStandardCodec.mm",
-    "framework/Source/FlutterStandardCodec_Internal.h",
     "framework/Source/FlutterTextInputDelegate.h",
     "framework/Source/FlutterTextInputPlugin.h",
     "framework/Source/FlutterTextInputPlugin.mm",
